+++ numpy/core/tests/test_blasdot.py
@@ -25,10 +25,9 @@ except ImportError:
 
 @dec.skipif(_dotblas is None, "Numpy is not compiled with _dotblas")
 def test_blasdot_used():
-    from numpy.core import dot, vdot, inner
-    assert_(dot is _dotblas.dot)
-    assert_(vdot is _dotblas.vdot)
-    assert_(inner is _dotblas.inner)
+    from numpy.core import vdot, inner
+    assert_(vdot is _dotblas.vdot, "vdot")
+    assert_(inner is _dotblas.inner, "inner")
 
 
 def test_dot_2args():
